File restructure for clarity
