--- app/src/main/java/com/example/lotto649/EventModel.java
@@ -1,4 +1,0 @@
-package com.example.lotto649;
-
-public class EventModel extends AbstractModel{
-}
