+++ app/src/main/java/com/example/lotto649/Models/EventModel.java
@@ -1,0 +1,6 @@
+package com.example.lotto649.Models;
+
+import com.example.lotto649.AbstractClasses.AbstractModel;
+
+public class EventModel extends AbstractModel {
+}
