--- app/src/main/java/com/example/lotto649/QrCodeModel.java
@@ -1,4 +1,0 @@
-package com.example.lotto649;
-
-public class QrCodeModel extends AbstractModel {
-}
