+++ app/src/main/java/com/example/lotto649/Views/Fragments/HomeFragment.java
@@ -9,13 +9,15 @@
  * <a href="https://www.geeksforgeeks.org/bottom-navigation-bar-in-android/">GeeksforGeeks: Bottom Navigation Bar in Android</a>
  * </p>
  */
-package com.example.lotto649;
+package com.example.lotto649.Views.Fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import androidx.fragment.app.Fragment;
+
+import com.example.lotto649.R;
 
 public class HomeFragment extends Fragment {
